refactor: Rename parser.go to loader.go, move loadRemoteFile
